Add bind_all support
Pusher has the functionality to subscribe to all events on a channel
or globally. This commit adds support for this functionality.

The use case is setting up an App level Pusher subscriptions
that then passes events on to angular $.on for improved scoping.

Pusher subscriptions have to be explicitly unsubscribed to otherwise
they will continue to exist and lead to buggy behavior.